fix: move dockertest config to own file
